requirements.txt
alembic==1.0.3
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
Click==7.0
colorama==0.4.0
coverage==4.5.2
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==1.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Mako==1.0.7
MarkupSafe==1.1.0
psycopg2==2.7.6.1
py==1.7.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.6.4
pytest==3.0.5
pytest-cov==2.4.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
simple-crypt==4.1.7
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
WTForms==2.2.1
gunicorn